MLAPI/requirements.txt
Flask
numpy
Pillow
keras
tensorflow-gpu==2.10.1
mysql-connector-python
